requirements.txt
GitPython
jupyter-book
mergedeep
pyyaml
